Fix incremental array updates
Fixes  incremental  array  updates  by  not  unwrapping   the   existing
observable  before  passing  it  into  the   (private)   updateViewModel
function.  updateViewModel  needs  an  observable  array  to  update  it
correctly, otherwise it will create a new empty array every time.

In the complex object case the  mappedRootObject  has  to  be  unwrapped
earlier to detect 'undefined' even if it is  wrapped  in  an  observable
(like  in  the  test:  "ko.mapping.fromJS  should  be  able  to   update
observables not created by fromJS"). Not sure if  this  breaks  'create'
and 'update' callbacks (OTOH it is not worse than  immediate  unwrapping
at the updateViewModel call site).